glim-backend/requirements.txt
Flask
flask_mysqldb
python-dotenv
mysql-connector-python
mysqlclient